requirements.txt
aiofiles==0.7.0
beautifulsoup4==4.9.3
blinker==1.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
docopt==0.6.2
expertai-nlapi==2.3.1
Flask==1.1.1
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
nltk==3.5
numpy==1.20.3
oauthlib==3.1.1
pandas==1.1.5
Pillow==8.2.0
pipreqs==0.4.10
plotly==4.14.3
priority==1.3.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rq==1.8.1
scipy==1.6.3
seaborn==0.10.1
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
tqdm==4.61.0
tweepy==3.10.0
urllib3==1.26.5
Werkzeug==2.0.1
wsproto==1.0.0
yarg==0.1.9
